third_party/blink/web_tests/dom/text/text-constructor-expected.txt

This tests that Text is constructable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new Text("one").data is "one"
PASS new Text().data is ""
PASS new Text(undefined).data is ""
PASS new Text(null).data is "null"
PASS new Text("two").ownerDocument is document
PASS typeof new Text is "object"
PASS Object.prototype.toString.call(new Text) is "[object Text]"
PASS new Text instanceof Text is true
PASS Object.getPrototypeOf(new Text) is Text.prototype
PASS new innerWindow.Text("three").ownerDocument is innerDocument
PASS new innerWindow.Text instanceof innerWindow.Text is true
PASS Object.getPrototypeOf(new innerWindow.Text) is innerWindow.Text.prototype
PASS successfullyParsed is true

TEST COMPLETE
